Move sexpr parsing into a new module
